Misc fixes for ossrecord/ossplay, FreeBSD compile, etc.
1) Fix FreeBSD compile - two typos by yours truly, and #ifdef around EBADE in setup/srcconf.c since that doesn't exist in FreeBSD. (EBADE is used only by libsalsa right now, and that doesn't compile in FreeBSD yet).
2) Change some nested if/else to switch/case in ossmix. Always print hexdecimal values with "0x" prefix.
3) Correct argument order in ossplay/seek_normal.
4) Remove not-yet-supported Xonar D2/D2X from .devices in oss_cmi878x.

Removed all *.[ch] files cmd/ossrecord/ , since ossplay directory should create ossrecord now.

+++ cmd/ossplay/ossplay.h
@@ -117,7 +117,8 @@ msadpcm_values_t;
 
 typedef ssize_t (decfunc_t) (unsigned char **, unsigned char *,
                              ssize_t, void *);
-typedef int (seekfunc_t) (int, unsigned int *, int, double, int, int);
+typedef ssize_t (seekfunc_t) (int, unsigned int *, unsigned int, double,
+                              unsigned int, int);
 
 typedef enum {
   FREE_OBUF = 1,
@@ -140,7 +141,7 @@ int le_int (const unsigned char *, int);
 off_t ossplay_lseek_stdin (int, off_t, int);
 int ossplay_parse_opts (int, char **, dspdev_t *);
 int ossrecord_parse_opts (int, char **, dspdev_t *);
-int play (dspdev_t *, int, unsigned int *, int, double,
+int play (dspdev_t *, int, unsigned int *, unsigned int, double,
           decoders_queue_t *, seekfunc_t *);
 int record (dspdev_t *, FILE *, const char *, double, unsigned int,
             unsigned int *, decoders_queue_t * dec);
